Move func to convert domain to legacy domain
This will put this in a common place where it can be used by other logic.
