+++ press/subscribers/purge_cache.py
@@ -1,16 +1,12 @@
 import requests
 
+from press.utils import convert_to_legacy_domain
+
 
 # Range of the sequence of ids to put into a purge url
 ID_CHUNK_SIZE = 10
 # Used by the quest
 TIMEOUT = (1, 5)  # (<connect>, <read>)
-
-
-def _build_legacy_domain(domain):
-    """Given the existing domain, translate it to the legacy domain."""
-    sep = len(domain.split('.')) > 2 and '-' or '.'
-    return 'legacy{}{}'.format(sep, domain)
 
 
 def _gen_purge_url(base, ids):
@@ -34,7 +30,7 @@ def purge_cache(event):
     just_ids = list(map(lambda x: x[0], event.ids))
 
     # Build the legacy domain from the current request domain.
-    domain = _build_legacy_domain(event.request.domain)
+    domain = convert_to_legacy_domain(event.request.domain)
     # Build the base part of the purge url
     scheme = event.request.scheme
     base_url = '{}://{}'.format(scheme, domain)
