+++ press/utils.py
@@ -1,4 +1,5 @@
 __all__ = (
+    'convert_to_legacy_domain',
     'convert_version_to_legacy_version',
 )
 
@@ -11,3 +12,9 @@ def convert_version_to_legacy_version(version):
 
     """
     return '1.{}'.format(version[0])
+
+
+def convert_to_legacy_domain(domain):
+    """Given the existing domain, convert it to the legacy domain."""
+    sep = len(domain.split('.')) > 2 and '-' or '.'
+    return 'legacy{}{}'.format(sep, domain)
