+++ tests/unit/test_utils.py
@@ -1,4 +1,7 @@
-from press.utils import convert_version_to_legacy_version
+from press.utils import (
+    convert_to_legacy_domain,
+    convert_version_to_legacy_version,
+)
 
 
 def test_module_version():
@@ -11,3 +14,16 @@ def test_collection_version():
     version = (8, 11)
     expected_version = '1.{}'.format(version[0])
     assert convert_version_to_legacy_version(version) == expected_version
+
+
+class TestConvertToLegacyDomain:
+
+    def test_prod_domain(self):
+        domain = 'example.com'
+        expected = '.'.join(['legacy', domain])
+        assert convert_to_legacy_domain(domain) == expected
+
+    def test_dev_domain(self):
+        domain = 'dev.example.com'
+        expected = '-'.join(['legacy', domain])
+        assert convert_to_legacy_domain(domain) == expected
